sprit train_config from train_predictor
