+++ diffusion_planner/diffusion_planner/train_config.py
@@ -1,0 +1,129 @@
+from dataclasses import dataclass, field
+from typing import Literal, Optional
+from diffusion_planner.dimensions import (
+    OUTPUT_T, INPUT_T, MAX_NUM_NEIGHBORS, 
+    NUM_SEGMENTS_IN_LANE, POINTS_PER_LANELET, 
+    NUM_SEGMENTS_IN_ROUTE, NUM_POLYGONS, POINTS_PER_POLYGON, 
+    NUM_LINE_STRINGS, POINTS_PER_LINE_STRING
+)
+
+from diffusion_planner.utils.normalizer import ObservationNormalizer, StateNormalizer
+
+
+@dataclass
+class TrainConfig:
+    # ---------------------------------------------------------
+    # Required Arguments (必須パラメータ: デフォルト値なしのものは先頭に置くルールです)
+    # ---------------------------------------------------------
+    exp_name: str
+    save_dir: str
+    train_set_list: str
+    valid_set_list: str
+
+    # ---------------------------------------------------------
+    # Data Dimensions
+    # ---------------------------------------------------------
+    future_len: int = OUTPUT_T
+    time_len: int = INPUT_T + 1
+    ego_prediction_horizon: int = OUTPUT_T
+
+    agent_state_dim: int = 11
+    agent_num: int = MAX_NUM_NEIGHBORS
+
+    static_objects_state_dim: int = 10
+    static_objects_num: int = 5
+
+    lane_num: int = NUM_SEGMENTS_IN_LANE
+    lane_len: int = POINTS_PER_LANELET
+
+    route_num: int = NUM_SEGMENTS_IN_ROUTE
+    route_len: int = POINTS_PER_LANELET
+
+    polygon_num: int = NUM_POLYGONS
+    polygon_len: int = POINTS_PER_POLYGON
+
+    line_string_num: int = NUM_LINE_STRINGS
+    line_string_len: int = POINTS_PER_LINE_STRING
+
+    # ---------------------------------------------------------
+    # DataLoader parameters
+    # ---------------------------------------------------------
+    use_data_augment: bool = True
+    augment_prob: float = 0.5
+    augment_type: Literal["quintic", "bridge"] = "quintic"
+    num_refine: int = 20
+    ego_past_noise_std: float = 0.1
+    use_smoothing_future_trajectory: bool = True
+    normalization_file_path: str = "normalization.json"
+    num_workers: int = 8
+    pin_mem: bool = True
+
+    # ---------------------------------------------------------
+    # Training
+    # ---------------------------------------------------------
+    seed: int = 3407
+    train_epochs: int = 100
+    batch_size: int = 512
+    save_utd: int = 10
+    learning_rate: float = 1e-4
+    warm_up_epoch: int = 5
+    encoder_drop_path_rate: float = 0.1
+    decoder_drop_path_rate: float = 0.1
+    use_ego_history: bool = True
+    ego_history_dropout_rate: float = 0.4
+    use_turn_indicators: bool = True
+
+    # Loss Coefficients
+    coeff_position_lat_loss: float = 1.0
+    coeff_position_lon_loss: float = 1.0
+    coeff_heading_l2_loss: float = 1.0
+    coeff_velocity: float = 1.0
+    # リストのようなミュータブル（変更可能）なデフォルト値は default_factory を使います
+    coeff_timestep: list[float] = field(default_factory=lambda: [1.0, 1.0, 1.0, 1.0])
+
+    coeff_road_border_loss: float = 1.0
+    road_border_margin: float = 0.25
+    road_border_n_interp: int = 2
+
+    coeff_neighbor_collision_loss: float = 0.0
+    neighbor_collision_margin: float = 0.25
+
+    alpha_planning_loss: float = 1.0
+    alpha_neighbor_loss: float = 0.1
+
+    # Velocity representation & hybrid loss
+    use_velocity_representation: bool = False
+    hybrid_loss_omega: float = 0.1
+    hybrid_loss_window: int = 10
+
+    guidance_scale: float = 0.5
+    device: str = "cuda"
+    use_ema: bool = True
+
+    # ---------------------------------------------------------
+    # Model Architecture
+    # ---------------------------------------------------------
+    encoder_mixer_depth: int = 6
+    encoder_fusion_depth: int = 6
+    decoder_depth: int = 3
+    num_heads: int = 8
+    hidden_dim: int = 256
+    diffusion_model_type: Literal["x_start", "flow_matching"] = "x_start"
+    predicted_neighbor_num: int = MAX_NUM_NEIGHBORS
+    resume_model_path: Optional[str] = None
+
+    # ---------------------------------------------------------
+    # Logging & Distributed
+    # ---------------------------------------------------------
+    use_wandb: bool = False
+    wandb_run_id: Optional[str] = None
+    wandb_project_name: str = "Diffusion-Planner"
+    notes: str = ""
+    ddp: bool = True
+    port: str = "22323"
+
+    # ---------------------------------------------------------
+    # Normalizers (学習実行時に初期化してセットするための枠)
+    # ---------------------------------------------------------
+    state_normalizer: Optional[StateNormalizer] = None
+    observation_normalizer: Optional[ObservationNormalizer] = None
